@@ -7,7 +7,7 @@ files                  = "src/**/*.py"
 ignore_errors          = true
 ignore_missing_imports = true
 pretty                 = true
-python_version         = "3.9"
+python_version         = "3.10"
 show_column_numbers    = true
 show_error_codes       = true
 show_error_context     = true
@@ -27,17 +27,17 @@ version     = "0.4.0"
 [tool.poetry.dependencies]
 bs4             = "*"
 html5lib        = "*"
-IPython         = "*"
+IPython         = "^8.4"
 kaleido         = "0.2.1"
-lxml            = "*"
-numpy           = "*"
-pandas          = "*"
-plotly          = "*"
-python          = "^3.9"
-python-dateutil = "*"
-pyyaml          = "*"
-requests        = "*"
-unidecode       = "*"
+lxml            = "^4.9"
+numpy           = "^1.23"
+pandas          = "^1.4"
+plotly          = "^5.10"
+python          = "^3.10"
+python-dateutil = "^2.8"
+pyyaml          = "^6.0"
+requests        = "^2.28"
+unidecode       = "^1.3"
 
 [tool.poetry.dev-dependencies]
 jinja2              = "<3.1.0"
@@ -52,7 +52,7 @@ pytest-cov          = "*"
 pytest-randomly     = "*"
 
 [tool.pyright]
-pythonVersion = "3.9"
+pythonVersion = "3.10"
 strictListInference = true
 
 [tool.pytest.ini_options]
