@@ -19,7 +19,9 @@ C1 = fab.Card(
     body             = '**Bravo Specialization** *(You may only have Crippling Crush in your deck if your hero is Bravo.)*\n\n**Crush** - If Crippling Crush deals 4 or more damage to a hero, they discard 2 random cards.',
     card_type        = 'Action',
     class_type       = 'Guardian',
+    color            = 'Red',
     cost             = 7,
+    dates            = {},
     defense          = 3,
     editions         = ['A', 'N', 'U'],
     effect_keywords  = ['Deal', 'Discard'],
@@ -28,7 +30,7 @@ C1 = fab.Card(
     full_name        = 'Crippling Crush (1)',
     grants_keywords  = [],
     identifiers      = ['1HP050', 'WTR043'],
-    image_urls       = ['https://storage.googleapis.com/fabmaster/media/images/1HP050.width-450.png', 'https://storage.googleapis.com/fabmaster/media/images/WTR_43.width-450.png', 'https://storage.googleapis.com/fabmaster/media/images/U-WTR43.width-450.png'],
+    image_urls       = {},
     intellect        = None,
     keywords         = ['Crush', 'Deal', 'Discard', 'Specialization'],
     label_keywords   = ['Crush'],
@@ -57,7 +59,9 @@ C2 = fab.Card(
     body             = 'If the next card you defend with this turn is a card with **combo**, it gains +2{d}.',
     card_type        = 'Defense Reaction',
     class_type       = 'Ninja',
+    color            = 'Yellow',
     cost             = 0,
+    dates            = {},
     defense          = 3,
     editions         = ['A', 'N', 'U'],
     effect_keywords  = [],
@@ -66,7 +70,7 @@ C2 = fab.Card(
     full_name        = 'Flic Flak (2)',
     grants_keywords  = [],
     identifiers      = ['1HP114', 'WTR093'],
-    image_urls       = ['https://storage.googleapis.com/fabmaster/media/images/1HP114.width-450.png', 'https://storage.googleapis.com/fabmaster/media/images/WTR_93.width-450.png', 'https://storage.googleapis.com/fabmaster/media/images/U-WTR93.width-450.png'],
+    image_urls       = {},
     intellect        = None,
     keywords         = ['Combo'],
     label_keywords   = ['Combo'],
@@ -95,18 +99,20 @@ C3 = fab.Card(
     body             = '**Once per Turn Action** - Create a Soul Shackle token: Your next Runeblade or Shadow action this turn gains **go again. Go again** *(It\'s a Shadow Runeblade aura with "At the beginning of your action phase, banish the top card of your deck.")*',
     card_type        = 'Hero',
     class_type       = 'Runeblade',
+    color            = None,
     cost             = None,
+    dates            = {},
     defense          = None,
     editions         = ['F', 'N', 'U'],
-    effect_keywords  = ['Create'],
+    effect_keywords  = ['Create', 'Gain'],
     flavor_text      = None,
     foilings         = ['S', 'R', 'C'],
     full_name        = 'Chane',
     grants_keywords  = ['Go again'],
     identifiers      = ['HER037', 'CHN001', 'MON154'],
-    image_urls       = ['https://storage.googleapis.com/fabmaster/media/images/HER037.width-450.png', 'https://storage.googleapis.com/fabmaster/media/images/CHN001.width-450.png', 'https://storage.googleapis.com/fabmaster/media/images/MON154.width-450.png', 'https://storage.googleapis.com/fabmaster/media/images/U-MON154.width-450.png'],
+    image_urls       = {},
     intellect        = 4,
-    keywords         = ['Action', 'Create', 'Go again', 'Soul Shackle'],
+    keywords         = ['Action', 'Create', 'Gain', 'Go again', 'Soul Shackle'],
     label_keywords   = [],
     legality         = {'B': True, 'C': True, 'CC': True},
     life             = 20,
