@@ -22,12 +22,12 @@ def test_collections():
     assert CL1.costs() == [0, 7]
     assert CL1.defense_values() == [3]
     assert CL1.editions() == ['A', 'F', 'N', 'U']
-    assert CL1.effect_keywords() == ['Create', 'Deal', 'Discard']
+    assert CL1.effect_keywords() == ['Create', 'Deal', 'Discard', 'Gain']
     assert CL1.foilings() == ['S', 'R', 'C']
     assert CL1.full_names() == ['Chane', 'Crippling Crush (1)', 'Flic Flak (2)']
     assert CL1.grants_keywords() == ['Go again']
     assert CL1.intellect_values() == [4]
-    assert CL1.keywords() == ['Action', 'Combo', 'Create', 'Crush', 'Deal', 'Discard', 'Go again', 'Soul Shackle', 'Specialization']
+    assert CL1.keywords() == ['Action', 'Combo', 'Create', 'Crush', 'Deal', 'Discard', 'Gain', 'Go again', 'Soul Shackle', 'Specialization']
     assert CL1.label_keywords() == ['Combo', 'Crush']
     assert CL1.life_values() == [20]
     assert CL1.names() == ['Chane', 'Crippling Crush', 'Flic Flak']
